upgrade buf and re-organise folder structure
