+++ buf.gen.yaml
@@ -1,0 +1,10 @@
+version: v2
+managed:
+  enabled: true
+  override:
+    - file_option: go_package_prefix
+      value: github.com/Azure/AgentBaker/pkg
+plugins:
+  - remote: buf.build/protocolbuffers/go
+    out: pkg
+    opt: paths=source_relative
