+++ pkg/aksnodeconfig/helpers/const.go
@@ -1,4 +1,4 @@
-package aksnodeconfigv1
+package helpers
 
 const (
 	contractVersion = "v0"
