+++ pkg/aksnodeconfig/helpers/utils_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-package aksnodeconfigv1
+package helpers
 
 import (
 	"reflect"
