--- pkg/proto/buf.gen.yaml
@@ -1,9 +1,0 @@
-version: v1
-managed:
-  enabled: true
-  go_package_prefix:
-    default: github.com/Azure/AgentBaker
-plugins:
-  - plugin: buf.build/protocolbuffers/go
-    out: ./
-    opt: paths=source_relative
