--- pkg/proto/buf.yaml
@@ -1,10 +1,0 @@
-version: v1
-breaking:
-  use:
-  - FILE
-lint:
-  use:
-  - DEFAULT
-  #- COMMENTS
-  - UNARY_RPC
-  - PACKAGE_NO_IMPORT_CYCLE
