+++ proto/aksnodeconfig/v1/authconfig.proto
