+++ proto/aksnodeconfig/v1/clusterconfig.proto
