+++ proto/aksnodeconfig/v1/containerdconfig.proto
