+++ proto/aksnodeconfig/v1/customcloudconfig.proto
