+++ proto/aksnodeconfig/v1/customsearchdomainconfig.proto
