+++ proto/aksnodeconfig/v1/gpuconfig.proto
