+++ proto/aksnodeconfig/v1/httpproxyconfig.proto
