+++ proto/aksnodeconfig/v1/imdsrestrictionconfig.proto
