+++ proto/aksnodeconfig/v1/kubebinaryconfig.proto
