+++ proto/aksnodeconfig/v1/kubeletconfig.proto
