+++ proto/aksnodeconfig/v1/networkconfig.proto
