+++ proto/aksnodeconfig/v1/teleportconfig.proto
