+++ proto/aksnodeconfig/v1/tlsbootstrappingconfig.proto
